Add message access protection using AspectJ
When a method of message, method of header of a message or method of service provided over a message is called from other thread than internal thread currently used for processing a message, an UnsupportedOperationException is thrown.

+ Add changes management for messages headers
Accessing header of a message is considered to be the same operation as calling method over the service. Getters and setter of headers are distinguished the same way as readonly and modifying methods of services..