Fix not committing changes in some situations
When an attempt to access a message is made, proxy commits only
changes made by service currently referenced by changedModuleBinding,
but not by services that would subsequently become referenced by
changedModuleBinding after calling doChanges on the 1st service.

Fix introduces committing changes in a loop until there's no service
referenced as only with up-to-date (changedModuleBinding) or it is
the service making an access attempt.